prefactoring: extract BaseCollector class
