+++ django_performance_testing/queries.py
@@ -1,13 +1,10 @@
-import copy
 import functools
-import traceback
 from django.db import connection
 from django.utils import six
-from django_performance_testing.signals import \
-    results_collected, before_clearing_queries_log
-from django_performance_testing.core import BaseLimit, LimitViolationError
+from django_performance_testing.signals import before_clearing_queries_log
+from django_performance_testing.core import \
+    BaseLimit, LimitViolationError, BaseCollector
 from django_performance_testing.utils import DelegatingProxy
-from django_performance_testing import context
 
 
 def setup_sending_before_clearing_queries_log_signal():
@@ -48,29 +45,9 @@ class QueryCountResult(object):
         return str(self.nr_of_queries)
 
 
-class QueryCollector(object):
+class QueryCollector(BaseCollector):
 
     _ids = set()
-
-    def __init__(self, id_=None):
-        self.id_ = id_
-        self.ensure_id_is_unique()
-
-    def ensure_id_is_unique(self):
-        if self.should_have_unique_id():
-            if self.id_ in self._ids:
-                id_ = self.id_
-                self.id_ = None
-                raise TypeError(
-                        'There is already a collector named {!r}'.format(id_))
-            self._ids.add(self.id_)
-
-    def __del__(self):
-        if self.should_have_unique_id():
-            self._ids.remove(self.id_)
-
-    def should_have_unique_id(self):
-        return self.id_ is not None
 
     def __enter__(self):
         self.queries = []
@@ -86,30 +63,11 @@ class QueryCollector(object):
         self.store_queries()
         self.nr_of_queries_when_entering = 0
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def before_exit(self):
         before_clearing_queries_log.disconnect(
             self.queries_about_to_be_reset_handler)
         connection.force_debug_cursor = self.orig_force_debug_cursor
         self.store_queries()
-        signal_responses = results_collected.send_robust(
-            sender=self, results=self.get_results_to_send(),
-            context=copy.deepcopy(context.current.data))
-        if exc_type is None:
-            for (receiver, response) in signal_responses:
-                if isinstance(response,  BaseException):
-                    orig_tb = ''.join(
-                        traceback.format_tb(response.__traceback__))
-                    error_msg = '{}{}: {}'.format(
-                        orig_tb,
-                        type(response).__name__,
-                        str(response)
-                    )
-                    if hasattr(response, 'clone_with_more_info'):
-                        new_exc = response.clone_with_more_info(
-                            orig_tb=orig_tb)
-                    else:
-                        new_exc = type(response)(error_msg)
-                    raise new_exc
 
     def get_results_to_send(self):
         by_type = dict(read=[], write=[], other=[])
